Move PGP package back to vault/helper
