+++ command/operator_init.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/vault/api"
-	"github.com/hashicorp/vault/sdk/helper/pgpkeys"
+	"github.com/hashicorp/vault/helper/pgpkeys"
 	"github.com/mitchellh/cli"
 	"github.com/posener/complete"
