+++ http/sys_rekey.go
@@ -8,8 +8,8 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/hashicorp/vault/helper/pgpkeys"
 	"github.com/hashicorp/vault/sdk/helper/consts"
-	"github.com/hashicorp/vault/sdk/helper/pgpkeys"
 	"github.com/hashicorp/vault/vault"
 )
